Missing base_path fix
This might break non-jplayer players, but IMHO it's the right thing:
drupal_add_js expects a path relative to base, so adding base_path()
to the string passed there feels wrong.

OTOH the swfPath variable of the jplayer flash thingy isn't mangled by
drupal, so it must be the final path.

Signed-off-by: Patrick Georgi <patrick@georgi-clan.de>